auth-service/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
pydantic[email]==2.11.10
python-dotenv==1.1.1
httpx==0.27.0
supabase==2.28.0
websockets==13.0
bcrypt==4.1.2
pyjwt==2.10.1